second large cleanup
- Removes some uneeded reflection and replaces with lambdas
- Changs package names
